pyproject.toml
[build-system]
requires = ["setuptools>=45"]
build-backend = "setuptools.build_meta"

[project]
name = "valet-studio"
version = "0.1.0"
description = "Valet Studio – deterministic audit + receipt + video pipeline"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.100",
  "uvicorn[standard]>=0.23",
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "pillow>=10.0",
  "moviepy>=1.0,<2.0",
  "numpy>=1.24",
  "trafilatura>=1.6",
  "yt-dlp>=2024.1",
  "openai-whisper>=20231117",
  "openai>=1.0",
  "anthropic>=0.20",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "httpx>=0.24",
  "black>=24.0",
  "ruff>=0.6",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
